refactor: move snapshot size calc to a new file and getNextBackup only return vol snap backup
