+++ pkg/backup/backup/backup_manager.go
@@ -508,7 +508,7 @@ func (bm *backupManager) makeBRBackupJob(backup *v1alpha1.Backup) (*batchv1.Job,
 			}
 		}
 	case v1alpha1.BackupModeVolumeSnapshot:
-		reason, err = bm.volSnapshotBackup(backup, tc)
+		reason, err = bm.volumeSnapshotBackup(backup, tc)
 		if err != nil {
 			return nil, reason, fmt.Errorf("backup %s/%s, %v", ns, name, err)
 		}
@@ -695,7 +695,7 @@ func (bm *backupManager) saveClusterMetaToExternalStorage(b *v1alpha1.Backup, cs
 	return "", nil
 }
 
-func (bm *backupManager) volSnapshotBackup(b *v1alpha1.Backup, tc *v1alpha1.TidbCluster) (string, error) {
+func (bm *backupManager) volumeSnapshotBackup(b *v1alpha1.Backup, tc *v1alpha1.TidbCluster) (string, error) {
 	if s, reason, err := snapshotter.NewSnapshotterForBackup(b.Spec.Mode, bm.deps); err != nil {
 		return reason, err
 	} else if s != nil {
